requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
asttokens 
async-timeout==4.0.3
attrs==23.2.0
autograd==1.6.2
autograd-gamma==0.5.0
ax-platform==0.3.7
botorch==0.10.0
brotlipy==0.7.0
CD==1.1.0
certifi 
cffi 
charset-normalizer 
chemprop==1.7.0
cloudpickle==3.0.0
comm 
contourpy==1.1.1
cryptography 
cycler==0.12.1
dataclasses==0.6
debugpy 
decorator 
DeepPurpose==0.1.5
descriptastorus 
dgl==2.0.0+cu118
dgllife==0.3.2
docstring_parser==0.16
exceptiongroup 
executing 
filelock 
fonttools==4.43.1
formulaic==1.0.1
frozenlist==1.4.1
fsspec==2023.9.2
future==1.0.0
fuzzywuzzy==0.18.0
gflownet 
gmpy2 
gpytorch==1.11
grpcio==1.63.0
huggingface-hub==0.21.4
hydra-core==1.3.2
hyperopt==0.2.7
idna 
importlib-metadata==6.8.0
importlib-resources==6.1.0
interface-meta==1.3.0
ipykernel 
ipython 
ipywidgets==8.1.2
jaxtyping==0.2.28
jedi 
Jinja2 
joblib==1.3.2
jupyter_client 
jupyter_core 
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
lifelines==0.28.0
lightning-utilities==0.11.0
linear-operator==0.5.1
Markdown==3.6
MarkupSafe 
matplotlib==3.8.0
matplotlib-inline 
mkl-fft 
mkl-random 
mkl-service==2.4.0
mock==5.1.0
mpmath 
multidict==6.0.5
multipledispatch==1.0.0
mypy-extensions==1.0.0
nest_asyncio 
networkx 
numpy 
omegaconf==2.3.0
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.1
pandas-flavor==0.6.0
parso 
pexpect 
pickleshare 
Pillow 
platformdirs 
plotly==5.22.0
prettytable==3.10.0
prompt-toolkit 
protobuf==5.26.0
psutil 
ptyprocess 
pure-eval 
py4j==0.10.9.7
pycparser 
pyg-lib==0.4.0+pt21cu118
Pygments 
pyOpenSSL 
pyparsing==3.1.1
pyre-extensions==0.0.30
pyro-api==0.1.2
pyro-ppl==1.9.0
PySocks 
PyTDC==0.4.1
python-dateutil==2.8.2
pytorch-lightning==2.2.1
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq 
rdkit==2023.9.5
rdkit-pypi==2022.9.5
requests 
scikit-learn==1.4.1.post1
scipy==1.11.3
seaborn==0.13.2
six 
stack-data 
subword-nmt==0.3.8
sympy 
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.3.0
torch==2.1.0
torch-cluster==1.6.3+pt21cu118
torch-scatter==2.1.2+pt21cu118
torch-sparse==0.6.18+pt21cu118
torch-spline-conv==1.2.2+pt21cu118
torch_geometric==2.5.0
torchaudio==2.1.0
torchmetrics==1.3.2
torchtyping==0.1.4
torchvision==0.16.0
tornado 
tqdm 
traitlets 
triton==2.1.0
typed-argument-parser==1.9.0
typeguard==2.13.3
typing-inspect==0.9.0
typing_extensions 
tzdata==2023.3
urllib3 
wcwidth 
Werkzeug==3.0.3
wget==3.2
widgetsnbextension==4.0.10
wrapt==1.16.0
xarray==2024.3.0
yarl==1.9.4
zipp 
